requirements.txt
asgiref==3.5.2
Django==4.0.6
sqlparse==0.4.2
tzdata==2022.1
psycopg2==2.9.3
django-admin-interface
Pillow==9.2.0
django-filter==22.1